Add support for ros.cmake packages
