+++ setup.cfg
@@ -41,6 +41,7 @@ include_package_data = true
 [options.entry_points]
 colcon_bundle.task.bundle =
     ros.catkin = colcon_ros_bundle.task.catkin.bundle:RosCatkinBundle
+    ros.cmake = colcon_ros_bundle.task.cmake.bundle:RosCmakeBundle
 colcon_bundle.installer =
     rdmanifest = colcon_ros_bundle.installer.rdmanifest:RdmanifestBundleInstallerExtensionPoint
